Fix for beefy

+++ defitrack-protocols/beefy/src/main/java/io/defitrack/protocol/beefy/domain/BeefyVault.kt
@@ -1,5 +1,7 @@
 package io.defitrack.protocol.beefy.domain
 
+import java.math.BigInteger
+
 class BeefyVault(
     val id: String,
     val name: String,
@@ -9,7 +11,7 @@ class BeefyVault(
     val earnedToken: String,
     val earnedTokenAddress: String,
     val earnContractAddress: String,
-    val pricePerFullShare: Int,
+    val pricePerFullShare: BigInteger,
     val status: String,
     val platform: String,
     val chain: String,
